refactor: move append tests to separate e2e file
Move append e2e tests from init-build.test.ts to dedicated append.test.ts
with expanded test coverage including:
- URL append with parameters
- CSV dialect options (delimiter, newLine, escapeChar)
- Format auto-detection (csv, ndjson, jsonl, parquet)
- Query string stripping
- Error handling (404, 429, 500)
- Validation errors
- Result structure verification
