+++ e2e/init-build.test.ts
@@ -741,178 +741,4 @@ describe("E2E: Init + Build Happy Path", () => {
       expect(allCalls).not.toContain("┌");
     });
   });
-
-  describe("datasources.append", () => {
-    it("appends data from URL using typed client", async () => {
-      let appendCalled = false;
-      let capturedDatasourceName: string | null = null;
-      let capturedMode: string | null = null;
-      let capturedFormat: string | null = null;
-      let capturedUrl: string | null = null;
-
-      server.use(
-        // Branch exists
-        http.get(`${BASE_URL}/v0/environments/:name`, () => {
-          return HttpResponse.json({
-            id: "branch-feature_test_branch",
-            name: "feature_test_branch",
-            token: "p.branch-token",
-            created_at: new Date().toISOString(),
-          });
-        }),
-        // Build endpoint
-        http.post(`${BASE_URL}/v1/build`, () => {
-          return HttpResponse.json(
-            createBuildSuccessResponse({
-              newDatasources: ["page_views"],
-              newPipes: ["top_pages"],
-            })
-          );
-        }),
-        // Append endpoint
-        http.post(`${BASE_URL}/v0/datasources`, async ({ request }) => {
-          appendCalled = true;
-          const url = new URL(request.url);
-          capturedDatasourceName = url.searchParams.get("name");
-          capturedMode = url.searchParams.get("mode");
-          capturedFormat = url.searchParams.get("format");
-          capturedUrl = (await request.text()).replace("url=", "");
-
-          return HttpResponse.json({
-            successful_rows: 100,
-            quarantined_rows: 0,
-            import_id: "import_123",
-          });
-        })
-      );
-
-      // Run init
-      await runInit({
-        cwd: tempDir,
-        skipLogin: true,
-        devMode: "branch",
-        clientPath: "lib/tinybird.ts",
-      });
-
-      // Run build to deploy resources
-      const buildResult = await runBuild({ cwd: tempDir });
-      expect(buildResult.success).toBe(true);
-
-      // Import the generated client and test append
-      const clientPath = path.join(tempDir, "lib/tinybird.ts");
-      const clientContent = fs.readFileSync(clientPath, "utf-8");
-
-      // The generated client exports `tinybird` - we need to use it
-      // Since we can't directly import the generated TS file, we'll test via the SDK directly
-      const { createClient } = await import("../src/client/base.js");
-
-      const client = createClient({
-        baseUrl: BASE_URL,
-        token: "p.branch-token",
-      });
-
-      // Test append from URL
-      const result = await client.datasources.append("page_views", {
-        url: "https://example.com/data.csv",
-      });
-
-      expect(appendCalled).toBe(true);
-      expect(capturedDatasourceName).toBe("page_views");
-      expect(capturedMode).toBe("append");
-      expect(capturedFormat).toBe("csv");
-      expect(decodeURIComponent(capturedUrl!)).toBe("https://example.com/data.csv");
-      expect(result.successful_rows).toBe(100);
-      expect(result.quarantined_rows).toBe(0);
-      expect(result.import_id).toBe("import_123");
-    });
-
-    it("appends data with CSV dialect options", async () => {
-      let capturedDelimiter: string | null = null;
-      let capturedNewLine: string | null = null;
-
-      server.use(
-        http.post(`${BASE_URL}/v0/datasources`, async ({ request }) => {
-          const url = new URL(request.url);
-          capturedDelimiter = url.searchParams.get("dialect_delimiter");
-          capturedNewLine = url.searchParams.get("dialect_new_line");
-
-          return HttpResponse.json({
-            successful_rows: 50,
-            quarantined_rows: 0,
-          });
-        })
-      );
-
-      const { createClient } = await import("../src/client/base.js");
-
-      const client = createClient({
-        baseUrl: BASE_URL,
-        token: "p.test-token",
-      });
-
-      await client.datasources.append("events", {
-        url: "https://example.com/data.csv",
-        csvDialect: {
-          delimiter: ";",
-          newLine: "\r\n",
-        },
-      });
-
-      expect(capturedDelimiter).toBe(";");
-      expect(capturedNewLine).toBe("\r\n");
-    });
-
-    it("auto-detects ndjson format from URL", async () => {
-      let capturedFormat: string | null = null;
-
-      server.use(
-        http.post(`${BASE_URL}/v0/datasources`, async ({ request }) => {
-          const url = new URL(request.url);
-          capturedFormat = url.searchParams.get("format");
-
-          return HttpResponse.json({
-            successful_rows: 25,
-            quarantined_rows: 0,
-          });
-        })
-      );
-
-      const { createClient } = await import("../src/client/base.js");
-
-      const client = createClient({
-        baseUrl: BASE_URL,
-        token: "p.test-token",
-      });
-
-      await client.datasources.append("events", {
-        url: "https://example.com/events.ndjson",
-      });
-
-      expect(capturedFormat).toBe("ndjson");
-    });
-
-    it("handles append API errors", async () => {
-      server.use(
-        http.post(`${BASE_URL}/v0/datasources`, () => {
-          return HttpResponse.json(
-            { error: "Datasource not found" },
-            { status: 404 }
-          );
-        })
-      );
-
-      const { createClient } = await import("../src/client/base.js");
-
-      const client = createClient({
-        baseUrl: BASE_URL,
-        token: "p.test-token",
-      });
-
-      await expect(
-        client.datasources.append("nonexistent", {
-          url: "https://example.com/data.csv",
-        })
-      ).rejects.toThrow("Datasource not found");
-    });
-  });
 });
